requirements.txt
fastapi
uvicorn
python-multipart
pythonocc-core